made tests more reliable

+++ src/DebugProject/ServiceWorker.cs
@@ -2,7 +2,9 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 using Topshelf;
@@ -12,6 +14,18 @@ namespace GainCapital.AutoUpdate.DebugProject
 	class ServiceWorker : ServiceControl
 	{
 	    private Process _updateProcess;
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        static extern bool GenerateConsoleCtrlEvent(UpdateChecker.ConsoleCtrlEvent sigevent, int dwProcessGroupId);
+        public enum ConsoleCtrlEvent
+        {
+            CTRL_C = 0,
+            CTRL_BREAK = 1,
+            CTRL_CLOSE = 2,
+            CTRL_LOGOFF = 5,
+            CTRL_SHUTDOWN = 6
+        }
+
 		public bool Start(HostControl hostControl)
 		{
 		    try
@@ -46,7 +60,24 @@ namespace GainCapital.AutoUpdate.DebugProject
 
 		public bool Stop(HostControl hostControl)
 		{
-            _updateProcess.Kill();
+		    try
+		    {
+		        if (!_updateProcess.WaitForExit(1))
+		        {
+                    Console.WriteLine("Sending CTRL+C to updater...");
+		            GenerateConsoleCtrlEvent(UpdateChecker.ConsoleCtrlEvent.CTRL_C, 0);
+		            if (!_updateProcess.WaitForExit(5000))
+		            {
+		                Console.WriteLine("Not shut down, killing...");
+		                _updateProcess.Kill();
+                        Thread.Sleep(5000);
+		            }
+		        }
+		    }
+		    catch
+		    {
+		    }
+            
 			_host = null;		
 			return true;
 		}
